+++ app/Http/Controllers/Admin/TeamController.php
@@ -10,17 +10,22 @@ class TeamController extends Controller
 {
     public function index()
     {
+        $this->authorize('view team');
+        
         $members = TeamMember::latest()->paginate(15);
         return view('admin.team.index', compact('members'));
     }
 
     public function create()
     {
+        $this->authorize('create team');
+        
         return view('admin.team.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create team');
         $request->validate([
             'name' => 'required',
             'email' => 'required|email|unique:team_members,email',
@@ -44,17 +49,20 @@ class TeamController extends Controller
         $data['slug'] = Str::slug($request->name);
 
         TeamMember::create($data);
-        return redirect()->route('team.index')->with('success', 'Team member added!');
+        return redirect()->route('team.index')->with('success', __('success.created', ['item' => __('Team Members')]));
     }
 
 
     public function edit(TeamMember $team)
     {
+        $this->authorize('edit team');
+        
         return view('admin.team.edit', compact('team'));
     }
 
     public function update(Request $request, TeamMember $team)
     {
+        $this->authorize('edit team');
         $request->validate([
             'name' => 'required',
             'email' => 'required|email|unique:team_members,email,' . $team->id,
@@ -84,18 +92,20 @@ class TeamController extends Controller
         $data['slug'] = Str::slug($request->name);
 
         $team->update($data);
-        return redirect()->route('team.index')->with('success', 'Team member updated!');
+        return redirect()->route('team.index')->with('success', __('success.updated', ['item' => __('Team Members')]));
     }
 
 
     public function destroy(TeamMember $team)
     {
+        $this->authorize('delete team');
+        
         $team->delete();
 
         if (request()->header('HX-Request')) {
             return response('', 200);
         }
 
-        return redirect()->route('team.index')->with('success', 'Team member deleted!');
+        return redirect()->route('team.index')->with('success', __('success.deleted', ['item' => __('Team Members')]));
     }
 }
